Fix REST API and logging
This fixes some bugs in the REST API, specifically around the get
projects functionality. This still needs to be updated to not just
dump everything in all projects and return a list that can be
iterated over to fetch elements for the projects, or at least
have multiple calls for when you don't need a dump of all
projects. This also follows some Actix best practices around
logging. This introduces bunyan formatted logging.