Update codeql and releases.yml
Codeql is updated to include go 1.21.X so it works right. This is
due to changes in the go.mod parsing in go > 1.21. By default
Codeql has Go 1.20 but I was informed that is just by coincidence
and the version of Go included should not be relied on.

Releases is updated with correct token permissions where needed.
Releases still doesn't work 100% yet and is being worked on.